+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ==================
 
   * Add /regdate command
-
+  * Add /define and /urbandefine commands
 
 1.5.0 / 2015-04-03
 ==================